Moved function and removed useless file
